requirements.txt
pip==21.3.1
setuptools==59.6.0
wheel==0.37.1
click
pyyaml